toggleThemeButton moved to useTheme
